comment class indentation error.

+++ main.py
@@ -141,15 +141,33 @@ class Article(db.Model):
 		if a :
 			return a
 
+	@classmethod
+	def check_if_user_owns_post(cls, post_id, uid):
+		a = cls.check_if_valid_post(post_id)
+		if a and a.created_by == uid.name:
+
 
 #### Comment model for database
+
 class Comment(db.Model):
 	created_by = db.StringProperty(required = True)
 	comment = db.TextProperty(required = True)
 	post_id = db.IntegerProperty(required = True)
 	created = db.DateTimeProperty(auto_now_add = True)
 	last_modified = db.DateTimeProperty(auto_now = True)
-		
+
+	@classmethod
+	def check_if_valid_comment(cls, comment_id):
+		key = db.Key.from_path('Comment', int(comment_id), parent=blog_key())
+		c = db.get(key)
+		if c :
+			return c
+	@classmethod
+	def check_if_user_owns_comment(cls, comment_id, uid):
+		c = cls.check_if_valid_comment(comment_id)
+		if c and c.created_by == uid.name
+
+
 #### Shows all the articles.
 class MainPage(Handler):
 	def get(self):
@@ -174,8 +192,7 @@ class PostPage(Handler):
 
 	#### Handle comments
 	def post(self, post_id):
-
-		a = check_if_valid_post(post_id)
+		a = Article.check_if_valid_post(post_id)
 		if a :
 			if self.user:
 				key = db.Key.from_path('Article', int(post_id), parent=blog_key())
@@ -217,24 +234,28 @@ class NewPost(Handler):
 		self.render("newpost.html")
 
 	def post(self):
-
-		if self.user:
-			title = self.request.get("title")
-			contents = self.request.get("contents")
-			created_by = self.user.name
-			
-			if title and contents:
-				a = Article(parent = blog_key(), title = title, 
-							contents = contents, created_by = created_by)
-				a.put()
-				self.redirect('/blog/%s' % str(a.key().id()))
+		a = Article.check_if_valid_post(post_id)
+		if a :
+			if self.user:
+				title = self.request.get("title")
+				contents = self.request.get("contents")
+				created_by = self.user.name
+				
+				if title and contents:
+					a = Article(parent = blog_key(), title = title, 
+								contents = contents, created_by = created_by)
+					a.put()
+					self.redirect('/blog/%s' % str(a.key().id()))
+				else:
+					error = "We need both a title and the blog content"
+					self.render("newpost.html", title = title, contents = contents,
+								 error = error)
 			else:
-				error = "We need both a title and the blog content"
-				self.render("newpost.html", title = title, contents = contents,
-							 error = error)
+				error = "Please login or register to write your story!"
+				self.render("error.html", error = error)
 		else:
-			error = "Please login or register to write your story!"
-			self.render("error.html", error = error)
+			self.error(404)
+			return
 
 
 #### Handles editing article.
@@ -246,20 +267,26 @@ class EditPost(Handler):
 
 
 	def post(self, post_id):
-		key = db.Key.from_path('Article', int(post_id), parent=blog_key())
-		a = db.get(key)
-		a.title = self.request.get("title")
-		a.contents = self.request.get("contents")
-		a.put()
-		self.redirect('/blog/%s' % post_id)
+		a = Article.check_if_valid_post(post_id)
+		if a and a.created_by == self.user.name:
+			a.title = self.request.get("title")
+			a.contents = self.request.get("contents")
+			a.put()
+			self.redirect('/blog/%s' % post_id)
+		else :
+			self.error(404)
+			return
 
 #### Deletes the article
 class DeletePost(Handler):
 	def get(self, post_id):
-		key = db.Key.from_path('Article', int(post_id), parent=blog_key())
-		a = db.get(key)
-		a.delete()
-		self.redirect('/blog/')
+		a = Article.check_if_user_owns_post(post_id, uid)
+		if a:
+			a.delete()
+			self.redirect('/blog/')
+		else :
+			error: "This is not your article"
+			self.render('error.html', error = error)
 
 #### Handles 'like' with the id in URL
 class LikeArticle(Handler):
@@ -308,28 +335,26 @@ class EditComment(Handler):
 			self.redirect("/blog/login")
 
 	def post(self, post_id, comment_id):
-		
-		if not self.user:
-			self.redirect('/blog/login')
 
-		comment = self.request.get('comment')
-
-		if comment:
-			key = db.Key.from_path('Comment', int(comment_id),
-									 parent=blog_key())
-			c = db.get(key)
-			c.comment = self.request.get('comment')
-			c.put()
-			self.redirect('/blog/%s' % post_id)
+		a = Article.check_if_valid_post(post_id)
+		if a :
+			if not self.user:
+				self.redirect('/blog/login')
+			c = Comment.check_if_user_owns_comment(comment_id, uid)
+			if c:
+				c.comment = self.request.get('comment')
+				c.put()
+				self.redirect('/blog/%s' % post_id)
+		else :
+			self.error(404)
+			return
 
 #### Delete comments based on article's id and comment's id
 class DeleteComment(Handler):
 	def get(self, post_id, comment_id):
 		if self.user:
-			key = db.Key.from_path('Comment', int(comment_id),
-									 parent=blog_key())
-			c = db.get(key)
-			if c.created_by == self.user.name:
+			c = Comment.check_if_user_owns_comment(comment_id, uid)
+			if c:
 				c.delete()
 				self.redirect('/blog/%s' % post_id)
 			else :
